fixed tables

+++ src/models/comment.ts
@@ -1,156 +1,6 @@
-// // src/models/Comment.ts
-// /**
-//  * MODELO COMMENT - SEQUELIZE-TYPESCRIPT CON DECORADORES
-//  */
-
-// import {
-//   Table,
-//   Column,
-//   Model,
-//   DataType,
-//   CreatedAt,
-//   UpdatedAt,
-//   DeletedAt,
-//   ForeignKey,
-//   BelongsTo,
-//   AllowNull,
-//   Length
-// } from 'sequelize-typescript';
-
-// // ⚠️ NO IMPORTAMOS los modelos relacionados aquí
-// // Las relaciones se definen con lazy loading
-
-// // ============================================
-// // INTERFACES
-// // ============================================
-
-// export interface CommentAttributes {
-//   id: number;
-//   post_id: number;
-//   user_id: number;
-//   content: string;
-//   created_at?: Date;
-//   updated_at?: Date;
-//   deleted_at?: Date | null;
-// }
-
-// // ============================================
-// // MODELO COMMENT CON DECORADORES
-// // ============================================
-
-// @Table({
-//   tableName: 'comments',
-//   timestamps: true,
-//   paranoid: true, // Soft delete
-//   underscored: true,
-//   charset: 'utf8mb4',
-//   collate: 'utf8mb4_unicode_ci',
-//   indexes: [
-//     {
-//       name: 'idx_comments_post_id',
-//       fields: ['post_id']
-//     },
-//     {
-//       name: 'idx_comments_user_id',
-//       fields: ['user_id']
-//     },
-//     {
-//       name: 'idx_comments_created_at',
-//       fields: ['created_at']
-//     }
-//   ]
-// })
-// export class Comment extends Model<CommentAttributes> {
-
-//   // ============================================
-//   // COLUMNAS
-//   // ============================================
-
-//   @Column({
-//     type: DataType.INTEGER.UNSIGNED,
-//     primaryKey: true,
-//     autoIncrement: true,
-//     comment: 'ID único del comentario'
-//   })
-//   declare id: number;
-
-//   @ForeignKey(() => require('./GobModelPost').default)
-//   @AllowNull(false)
-//   @Column({
-//     type: DataType.INTEGER.UNSIGNED,
-//     comment: 'ID del post al que pertenece'
-//   })
-//   declare post_id: number;
-
-//   @ForeignKey(() => require('./User').User)
-//   @AllowNull(false)
-//   @Column({
-//     type: DataType.INTEGER.UNSIGNED,
-//     comment: 'ID del usuario autor del comentario'
-//   })
-//   declare user_id: number;
-
-//   @Length({
-//     min: 1,
-//     max: 5000,
-//     msg: 'El comentario debe tener entre 1 y 5000 caracteres'
-//   })
-//   @AllowNull(false)
-//   @Column({
-//     type: DataType.TEXT,
-//     comment: 'Contenido del comentario'
-//   })
-//   declare content: string;
-
-//   // ============================================
-//   // TIMESTAMPS
-//   // ============================================
-
-//   @CreatedAt
-//   @Column({
-//     type: DataType.DATE,
-//     field: 'created_at'
-//   })
-//   declare created_at: Date;
-
-//   @UpdatedAt
-//   @Column({
-//     type: DataType.DATE,
-//     field: 'updated_at'
-//   })
-//   declare updated_at: Date;
-
-//   @DeletedAt
-//   @Column({
-//     type: DataType.DATE,
-//     field: 'deleted_at'
-//   })
-//   declare deleted_at: Date | null;
-
-//   // ============================================
-//   // RELACIONES (LAZY LOADING)
-//   // ============================================
-
-//   // Cada comentario pertenece a un post
-//   @BelongsTo(() => require('./GobModelPost').default, {
-//     foreignKey: 'post_id',
-//     as: 'post'
-//   })
-//   declare post?: any;
-
-//   // Cada comentario pertenece a un usuario
-//   @BelongsTo(() => require('./User').User, {
-//     foreignKey: 'user_id',
-//     as: 'author'
-//   })
-//   declare author?: any;
-// }
-
-// export default Comment;
 // src/models/Comment.ts
 /**
  * MODELO COMMENT - Sequelize-TypeScript con Decoradores
- * Versión híbrida respetando el estilo de tu compañera
  */
 
 import 'reflect-metadata';
@@ -218,19 +68,19 @@ export class Comment extends Model<CommentAttributes, CommentCreationAttributes>
     @SqComment('Identificador único del comentario')
     @PrimaryKey
     @AutoIncrement
-    @Column(DataType.INTEGER.UNSIGNED)
+    @Column(DataType.BIGINT.UNSIGNED) // 👈 CAMBIADO de INTEGER a BIGINT
     public id!: number;
 
     @SqComment('ID del post al que pertenece el comentario')
     @AllowNull(false)
     @ForeignKey(() => require('./GobModelPost').default)
-    @Column(DataType.INTEGER.UNSIGNED)
+    @Column(DataType.BIGINT.UNSIGNED) // 👈 CAMBIADO de INTEGER a BIGINT
     public post_id!: number;
 
     @SqComment('ID del usuario que hizo el comentario')
     @AllowNull(false)
     @ForeignKey(() => require('./User').User)
-    @Column(DataType.INTEGER.UNSIGNED)
+    @Column(DataType.BIGINT.UNSIGNED) // 👈 CAMBIADO de INTEGER a BIGINT
     public user_id!: number;
 
     @SqComment('Contenido del comentario')
